+++ app/frontend/Index.php
@@ -81,7 +81,7 @@ class Index extends Frontend
             'article_type_options'=>$articleTypeOptions,
             'homepage_research_cards' => ArticleModel::getHomepageFeaturedArticles('research', 2),
             'homepage_fragment_cards' => ArticleModel::getHomepageFeaturedArticles('fragment', 2),
-            'archive_chapters'=>HelpModel::getHomepageArchiveHighlights(4, 3),
+            'archive_chapters'=>HelpModel::getHomepageArchiveHighlights(2, 3),
             'knowledge_map_summary' => HelpModel::getKnowledgeMapSummary(),
             'knowledge_map_connections' => HelpModel::getKnowledgeMapTopicConnections(3, 2),
         ]);